2022-05-10, preccision error was the problem. It could be tested changing the vtype from double to long double

+++ source/output_main_parameters.c
@@ -42,7 +42,7 @@ void output_main_parameters(vtype final_time, int Number_timesteps, int Number_o
     fprintf(file, "lmin %d\n", lmin);
     fprintf(file, "lmax %d\n", lmax);
     fprintf(file, "_MAX_NUMBER_OF_ITERATIONS_IN_POISSON_EQUATION_ %d\n", _MAX_NUMBER_OF_ITERATIONS_IN_POISSON_EQUATION_);
-    fprintf(file, "_ERROR_THRESHOLD_IN_THE_POISSON_EQUATION_ %1.5e\n", _ERROR_THRESHOLD_IN_THE_POISSON_EQUATION_);
+    fprintf(file, "_ERROR_THRESHOLD_IN_THE_POISSON_EQUATION_ %1.5e\n", (double) _ERROR_THRESHOLD_IN_THE_POISSON_EQUATION_);
     fprintf(file, "check_poisson_error_method %d\n", check_poisson_error_method);
     fprintf(file, "multigrid_method %d\n", multigrid);
     fprintf(file, "solver %d\n", solver);
@@ -50,9 +50,9 @@ void output_main_parameters(vtype final_time, int Number_timesteps, int Number_o
     fprintf(file, "_NiterPostS_ %d\n", _NiterPostS_);
     fprintf(file, "_Niterfinddphic_ %d\n", _Niterfinddphic_);
     fprintf(file, "_Iter_branches_SOR_ %d\n", _Iter_branches_SOR_);
-    fprintf(file, "_w_SOR_ %1.5e\n", _w_SOR_);
-    fprintf(file, "_CFL_ %1.5e\n", _CFL_);
-    fprintf(file, "_MAX_dt_ %1.5e\n", _MAX_dt_);
+    fprintf(file, "_w_SOR_ %1.5e\n", (double) _w_SOR_);
+    fprintf(file, "_CFL_ %1.5e\n", (double) _CFL_);
+    fprintf(file, "_MAX_dt_ %1.5e\n",(double) _MAX_dt_);
 
     fprintf(file, "Ngrid %d\n", no_lmin_cell + 1);
     fprintf(file, "Nparticles %d\n", GL_no_ptcl);
